Move function to related file
