+++ src/esssans/normalization.py
@@ -28,6 +28,8 @@ from .types import (
     TransmissionFraction,
     TransmissionRun,
     UncertaintyBroadcastMode,
+    WavelengthBands,
+    WavelengthBins,
 )
 from .uncertainty import (
     broadcast_to_events_with_upper_bound_variances,
@@ -298,6 +300,42 @@ def iofq_denominator(
     return CleanWavelength[RunType, Denominator](denominator)
 
 
+def process_wavelength_bands(
+    wavelength_bands: Optional[WavelengthBands],
+    wavelength_bins: WavelengthBins,
+) -> ProcessedWavelengthBands:
+    """
+    Perform some checks and potential reshaping on the wavelength bands.
+
+    The wavelength bands must be either one- or two-dimensional.
+    If the wavelength bands are defined as a one-dimensional array, convert them to a
+    two-dimensional array with start and end wavelengths.
+
+    The final bands must have a size of 2 in the wavelength dimension, defining a start
+    and an end wavelength.
+    """
+    if wavelength_bands is None:
+        wavelength_bands = sc.concat(
+            [wavelength_bins.min(), wavelength_bins.max()], dim='wavelength'
+        )
+    if wavelength_bands.ndim == 1:
+        wavelength_bands = sc.concat(
+            [wavelength_bands[:-1], wavelength_bands[1:]], dim='x'
+        ).rename(x='wavelength', wavelength='band')
+    if wavelength_bands.ndim != 2:
+        raise ValueError(
+            'Wavelength_bands must be one- or two-dimensional, '
+            f'got {wavelength_bands.ndim}.'
+        )
+    if wavelength_bands.sizes['wavelength'] != 2:
+        raise ValueError(
+            'Wavelength_bands must have a size of 2 in the wavelength dimension, '
+            'defining a start and an end wavelength, '
+            f'got {wavelength_bands.sizes["wavelength"]}.'
+        )
+    return wavelength_bands
+
+
 def normalize(
     numerator: CleanSummedQ[RunType, Numerator],
     denominator: CleanSummedQ[RunType, Denominator],
@@ -380,5 +418,6 @@ providers = (
     iofq_norm_wavelength_term,
     iofq_denominator,
     normalize,
+    process_wavelength_bands,
     solid_angle,
 )
